requirements.txt
bs4
playwright==1.57.0
python-dotenv==1.2.1
playwright-stealth==2.0.0
pyserial==3.5
requests==2.32.5
python-dateutil==2.8.2
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
google-cloud-pubsub>=2.0.0
memory-profiler
uvicorn